+++ brownie_tokens/__init__.py
@@ -1,2 +1,6 @@
-from brownie_tokens.forked import MintableForkToken
+import brownie
+
+from brownie_tokens.forked import BrownieTokensError, MintableForkToken, skip_holders
 from brownie_tokens.template import ERC20
+
+brownie.config["autofetch_sources"] = True