+++ brownie_tokens/forked.py
@@ -1,12 +1,28 @@
+import os
 import requests
 import sys
-from brownie import Contract, Wei
+from brownie import Contract, Wei, web3
 from brownie.convert import to_address
-from typing import Dict, List
+from typing import Dict, List, Optional
+
+_ethplorer_api_key: str = os.getenv("ETHPLORER_API_KEY") or "freekey"
 
 _token_holders: Dict = {}
+_skip_list: List[str] = []
 
 _token_names = ["Aave"]
+
+
+class BrownieTokensError(Exception):
+    pass
+
+
+def update_skipped_addresses(token: Optional[str] = None) -> None:
+    if token:
+        _token_holders[token] = [a for a in _token_holders[token] if a not in _skip_list]
+    else:
+        for token in _token_holders:
+            update_skipped_addresses(token)
 
 
 def get_top_holders(address: str) -> List:
@@ -14,14 +30,36 @@ def get_top_holders(address: str) -> List:
     if address not in _token_holders:
         holders = requests.get(
             f"https://api.ethplorer.io/getTopTokenHolders/{address}",
-            params={"apiKey": "freekey", "limit": "50"},
+            params={"apiKey": _ethplorer_api_key, "limit": "50"},
         ).json()
+        if "error" in holders:
+            api_key_message = ""
+            if _ethplorer_api_key == "freekey":
+                api_key_message = (
+                    " Adding $ETHPLORER_API_KEY (from https://ethplorer.io) as environment variable"
+                    " may solve the problem."
+                )
+            raise BrownieTokensError(
+                f"Ethplorer returned error: {holders['error']}." + api_key_message
+            )
+
         _token_holders[address] = [to_address(i["address"]) for i in holders["holders"]]
         if address in _token_holders[address]:
             # don't steal from the treasury - that could cause wierdness
             _token_holders[address].remove(address)
+        update_skipped_addresses(address)
 
     return _token_holders[address]
+
+
+def skip_holders(*addresses: str) -> None:
+    """
+    Addresses that will remain untouched via '_mint_for_testing'.
+    Does not include minting with custom logic.
+    """
+    global _skip_list
+    _skip_list = list(set(_skip_list + list(addresses)))
+    update_skipped_addresses()
 
 
 class MintableForkToken(Contract):
@@ -101,7 +139,7 @@ def mint_0xdB25f211AB05b1c97D595516F45794528a807ad8(
     token: MintableForkToken, target: str, amount: int
 ) -> None:
     # EURS
-    owner = "0x2EbBbc541E8f8F24386FA319c79CedA0579f1Efb"
+    owner = web3.eth.getStorageAt(token.address, 2)[-20:].hex()
     token.createTokens(amount, {"from": owner})
     token.transfer(target, amount, {"from": owner})